cubeclassifyer/requirements.txt
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.7.0
numpy>=1.24.0
Pillow>=9.5.0
